feat(audit): refactor audit trail service to use protocol interface and DynamoDB implementation
